add tel searchrecent and tel delrecent commands
these look through all discord channels bridged to the given channel
(including that channel itself) and search for messages matching
the given query. the delrecent command, after asking for confirmation,
will then delete all of those messages that it can across all the channels.
it is capable of deleting messages sent by our own webhooks even if
we don't have manage message permissions in the channel in question.

the user interface is not amazing but these are both extpriv-only and
only send any outputs in dms so this probably isn't very important.